+++ leeger/model/stat/YearStatSheet.py
@@ -55,6 +55,7 @@ class YearStatSheet:
 
     # Optional Stats
     leagueMedianWins: Optional[dict[str, float]] = None
+    opponentLeagueMedianWins: Optional[dict[str, float]] = None
     totalGames: Optional[dict[str, int]] = None
 
     def preferredOrderWithTitle(self) -> list[tuple[str, dict]]:
@@ -96,5 +97,6 @@ class YearStatSheet:
         if self.leagueMedianWins is not None:
             response.insert(1, ("Total Games", self.totalGames))
             response.insert(4, ("League Median Wins", self.leagueMedianWins))
+            response.insert(5, ("Opponent League Median Wins", self.opponentLeagueMedianWins))
 
         return response