+++ leeger/util/stat_sheet.py
@@ -81,10 +81,12 @@ def leagueStatSheet(league: League, **kwargs) -> AllTimeStatSheet:
     # check for optional stats
     totalGames = None
     leagueMedianWins = None
+    opponentLeagueMedianWins = None
     for year in league.years:
         if year.yearSettings.leagueMedianGames is True:
             totalGames = TeamSummaryAllTimeCalculator.getTotalGames(league, **kwargs)
             leagueMedianWins = GameOutcomeAllTimeCalculator.getLeagueMedianWins(league, **kwargs)
+            opponentLeagueMedianWins = GameOutcomeAllTimeCalculator.getOpponentLeagueMedianWins(league, **kwargs)
             break
 
     return AllTimeStatSheet(gamesPlayed=gamesPlayed, wins=wins, losses=losses, ties=ties, winPercentage=winPercentage,
@@ -97,7 +99,8 @@ def leagueStatSheet(league: League, **kwargs) -> AllTimeStatSheet:
                             opponentScoringShare=opponentScoringShare, maxScore=maxScore, minScore=minScore,
                             scoringStandardDeviation=scoringStandardDeviation, plusMinus=plusMinus,
                             adjustedTeamScore=adjustedTeamScore, adjustedTeamSuccess=adjustedTeamSuccess,
-                            adjustedTeamLuck=adjustedTeamLuck, leagueMedianWins=leagueMedianWins, totalGames=totalGames)
+                            adjustedTeamLuck=adjustedTeamLuck, leagueMedianWins=leagueMedianWins, totalGames=totalGames,
+                            opponentLeagueMedianWins=opponentLeagueMedianWins)
 
 
 def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
@@ -151,9 +154,11 @@ def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
     # check for optional stats
     totalGames = None
     leagueMedianWins = None
+    opponentLeagueMedianWins = None
     if year.yearSettings.leagueMedianGames is True:
         totalGames = TeamSummaryYearCalculator.getTotalGames(year, **kwargs)
         leagueMedianWins = GameOutcomeYearCalculator.getLeagueMedianWins(year, **kwargs)
+        opponentLeagueMedianWins = GameOutcomeYearCalculator.getOpponentLeagueMedianWins(year, **kwargs)
 
     return YearStatSheet(gamesPlayed=gamesPlayed, wins=wins, losses=losses, ties=ties, winPercentage=winPercentage,
                          wal=wal, walPerGame=walPerGame, awal=awal, awalPerGame=awalPerGame, opponentAWAL=opponentAWAL,
@@ -165,4 +170,5 @@ def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
                          opponentScoringShare=opponentScoringShare, maxScore=maxScore, minScore=minScore,
                          scoringStandardDeviation=scoringStandardDeviation, plusMinus=plusMinus,
                          teamScore=teamScore, teamSuccess=teamSuccess, teamLuck=teamLuck,
-                         leagueMedianWins=leagueMedianWins, totalGames=totalGames)
+                         leagueMedianWins=leagueMedianWins, totalGames=totalGames,
+                         opponentLeagueMedianWins=opponentLeagueMedianWins)